#1: Implement an angular application to consume the soap service
Add jQuery and refactor illustrator component

Added jQuery and its types to the dependencies. Most importantly, the illustrator component was heavily refactored to include an implementation of an Expandable Panel with reusable attribute templates for a GoJS diagram. The GoJS dependency was also added in the type definitions in the tsconfig file.