Add development environment and enhance illustrator component
The development environment configuration was defined in the environment file. The illustrator component was enhanced with a new DbObjectsInfoService, foreign key handling and DbSchemas selection was added. The AppModule was created and contains the HttpClientModule and the DbObjectsInfoService as providers. Small changes in CSS file were made to match new layout requirements.